+++ src/main/java/sk/stuba/fei/uim/oop/assignment3/product/service/IProductService.java
@@ -9,9 +9,9 @@ import java.util.List;
 public interface IProductService {
         List<Product> getList();
         Product create(ProductRequest product);
-        Product getRetrieve(int id);
-        Product update(ProductRequest request, int id);
-        void delete(int id);
-        Amount getAmount(int id);
-        Amount addAmount(ProductRequest request, int id);
+        Product getRetrieve(Long id);
+        Product update(ProductRequest request, Long id);
+        void delete(Long id);
+        Amount getAmount(Long id);
+        Amount addAmount(ProductRequest request, Long id);
 }
